♻️ pull state setter up & move all utils to own folder
