+++ dashboard-app/src/components/StackedBarChart/Chart.tsx
@@ -5,7 +5,7 @@ import { Bar, ChartData } from 'react-chartjs-2';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 import * as chartjs from 'chart.js';
 
-import { getStackedGraphOptions, totalizer } from './util';
+import { getStackedGraphOptions, totalizer } from './utils/chartjsUtils';
 import _Card from '../Card';
 import { H3 } from '../Headings';
 import { ColoursEnum } from '../../styles/design_system';
