+++ dashboard-app/src/components/StackedBarChart/utils/chartjsUtils.ts
